+++ src/test/fuzz/package_eval.cpp
@@ -340,6 +340,8 @@ FUZZ_TARGET(ephemeral_package_eval, .init = initialize_tx_pool)
 
     node.validation_signals->UnregisterSharedValidationInterface(outpoints_updater);
 
+    chainstate.SetMempool(g_setup->m_node.mempool.get());
+
     WITH_LOCK(::cs_main, tx_pool.check(chainstate.CoinsTip(), chainstate.m_chain.Height() + 1));
 }
 
@@ -534,6 +536,8 @@ FUZZ_TARGET(tx_package_eval, .init = initialize_tx_pool)
     }
 
     node.validation_signals->UnregisterSharedValidationInterface(outpoints_updater);
+
+    chainstate.SetMempool(g_setup->m_node.mempool.get());
 
     WITH_LOCK(::cs_main, tx_pool.check(chainstate.CoinsTip(), chainstate.m_chain.Height() + 1));
 }
