+++ src/test/fuzz/validation_load_mempool.cpp
@@ -4,22 +4,21 @@
 
 #include <node/mempool_persist.h>
 
-#include <node/mempool_args.h>
 #include <node/mempool_persist_args.h>
 #include <test/fuzz/FuzzedDataProvider.h>
 #include <test/fuzz/fuzz.h>
 #include <test/fuzz/util.h>
-#include <test/fuzz/util/mempool.h>
 #include <test/util/setup_common.h>
-#include <test/util/txmempool.h>
 #include <txmempool.h>
-#include <util/check.h>
 #include <util/time.h>
-#include <util/translation.h>
 #include <validation.h>
 
-#include <cstdint>
-#include <vector>
+#include <functional>
+#include <memory>
+
+namespace fs {
+class path;
+}
 
 using node::DumpMempool;
 using node::LoadMempool;
@@ -27,12 +26,12 @@ using node::LoadMempool;
 using node::MempoolPath;
 
 namespace {
-const TestingSetup* g_setup;
+TestingSetup* g_setup;
 } // namespace
 
 void initialize_validation_load_mempool()
 {
-    static const auto testing_setup = MakeNoLogFileContext<const TestingSetup>();
+    static auto testing_setup = MakeNoLogFileContext<TestingSetup>();
     g_setup = testing_setup.get();
 }
 
@@ -43,12 +42,9 @@ FUZZ_TARGET(validation_load_mempool, .init = initialize_validation_load_mempool)
     SetMockTime(ConsumeTime(fuzzed_data_provider));
     FuzzedFileProvider fuzzed_file_provider{fuzzed_data_provider};
 
-    bilingual_str error;
-    CTxMemPool pool{MemPoolOptionsForTest(g_setup->m_node), error};
-    Assert(error.empty());
+    auto& pool{g_setup->ReplaceMempool()};
 
-    auto& chainstate{static_cast<DummyChainState&>(g_setup->m_node.chainman->ActiveChainstate())};
-    chainstate.SetMempool(&pool);
+    auto& chainstate{g_setup->m_node.chainman->ActiveChainstate()};
 
     auto fuzzed_fopen = [&](const fs::path&, const char*) {
         return fuzzed_file_provider.open();
