+++ src/test/miner_tests.cpp
@@ -51,19 +51,7 @@ struct MinerTestingSetup : public TestingSetup {
     }
     CTxMemPool& MakeMempool()
     {
-        // Delete the previous mempool to ensure with valgrind that the old
-        // pointer is not accessed, when the new one should be accessed
-        // instead.
-        m_node.mempool.reset();
-        bilingual_str error;
-        auto opts = MemPoolOptionsForTest(m_node);
-        // The "block size > limit" test creates a cluster of 1192590 vbytes,
-        // so set the cluster vbytes limit big enough so that the txgraph
-        // doesn't become oversized.
-        opts.limits.cluster_size_vbytes = 1'200'000;
-        m_node.mempool = std::make_unique<CTxMemPool>(opts, error);
-        Assert(error.empty());
-        return *m_node.mempool;
+        return ReplaceMempool();
     }
     std::unique_ptr<Mining> MakeMining()
     {
